+++ src/vmware/azext_vmware/aaz/latest/vmware/private_cloud/_rotate_vcenter_password.py
@@ -23,9 +23,9 @@ class RotateVcenterPassword(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2022-05-01",
+        "version": "2023-03-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.avs/privateclouds/{}/rotatevcenterpassword", "2022-05-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.avs/privateclouds/{}/rotatevcenterpassword", "2023-03-01"],
         ]
     }
 
@@ -51,6 +51,9 @@ class RotateVcenterPassword(AAZCommand):
             help="Name of the private cloud",
             required=True,
             id_part="name",
+            fmt=AAZStrArgFormat(
+                pattern="^[-\w\._]+$",
+            ),
         )
         _args_schema.resource_group = AAZResourceGroupNameArg(
             required=True,
@@ -134,7 +137,7 @@ class RotateVcenterPassword(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2022-05-01",
+                    "api-version", "2023-03-01",
                     required=True,
                 ),
             }