+++ src/connectedk8s/HISTORY.rst
@@ -2,6 +2,9 @@
 
 Release History
 ===============
+1.4.2
+++++++
+* Fix reference error.
 
 1.4.1
 ++++++